"tryy"

+++ 1/Hash_fun.h
@@ -9,10 +9,10 @@ using namespace std;
 const int mod_adler= 65521;
 class Hash_fun{
 public:
-    static int adler32(ifstream& file);
-    static int sum64(ifstream& file);
+    static uint32_t adler32(ifstream& file);
+    static uint64_t sum64(ifstream& file);
 private:
-    static int read_int(std::istream &file);
+    static uint64_t read_int(std::istream &file);
 };
